upgraded to newest ng2 packages improved test config
