+++ src/index.html
@@ -1,6 +1,5 @@
 <!doctype html>
 <html>
-
 <head>
   <meta charset="utf-8">
   <title>Web</title>
@@ -40,12 +39,10 @@
     }
   </style>
 </head>
-
 <body>
   <app-root></app-root>
-  <div class="loading">
+    <div class="loading">
     <h1><img alt="Brand" class="logo" src="assets/logo.png"> Loading...</h1>
   </div>
 </body>
-
 </html>
