+++ src/typings.d.ts
@@ -1,4 +1,2 @@
 // Typings reference file, you can add your own global typings here
 // https://www.typescriptlang.org/docs/handbook/writing-declaration-files.html
-
-declare var System: any;
